% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{extent.py}
\alias{extent.py}
\alias{extent.py.rasterio._io.DatasetReaderBase}
\title{Title}
\usage{
extent.py(x, ...)

\method{extent.py}{rasterio._io.DatasetReaderBase}(x, ...)
}
\arguments{
\item{x}{object, opened dataset or WarpedVRT dataset}

\item{...}{ignored}
}
\value{
xmin,xmax,ymin,ymax
}
\description{
Title
}
\examples{
extent.py(open.py(rgba_tif()))
}
