% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open.py}
\alias{open.py}
\title{Title}
\usage{
\method{open}{py}(
  fp,
  mode = "r",
  driver = NULL,
  width = NULL,
  height = NULL,
  count = NULL,
  crs = NULL,
  transform = NULL,
  dtype = NULL,
  nodata = NULL,
  sharing = FALSE
)
}
\arguments{
\item{fp}{file path, or url, or database connection string, or VRT string, anything GDAL can open as raster}

\item{mode}{open mode, readonly by default}

\item{driver}{driver to use, optional is autodetected}

\item{width}{width of data set (ignored in readonly)}

\item{height}{height of data set (ignored in readonly)}

\item{count}{number of bands (ignored in readonly)}

\item{crs}{projection of data set (ignored in readonly)}

\item{transform}{geotransform of data set, see details (ignored in readonly)}

\item{dtype}{data type of dataset (ignored in readonly)}

\item{nodata}{missing data value of dataset (ignored in readonly)}

\item{sharing}{open shared mode (FALSE by default)}
}
\value{

}
\description{
Title
}
\examples{
x <- open.py(rgba_tif())
}
