% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{plot.rasterio._io.DatasetReaderBase}
\alias{plot.rasterio._io.DatasetReaderBase}
\title{Title}
\usage{
\method{plot}{rasterio._io.DatasetReaderBase}(x, ..., dimension = if (dev.cur() == 1L) c(256, 256) else dev.size("px"))
}
\arguments{
\item{x}{open dataset or WarpedVRT dataset}

\item{dimension}{size of array to plot (determine automatically, use \code{NULL} for native)}
}
\value{
used for side effect, a plot
}
\description{
We can't specify bands yet.
}
\examples{
plot(open.py(rgba_tif()))
}
