% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{rgba_tif}
\alias{rgba_tif}
\title{Example RGBA GeoTIFF image file}
\usage{
rgba_tif()
}
\value{

}
\description{
Function returns a path to a file, for examples.
}
\details{
The GeoTIFF is the canonical example file, used by rasterio examples. This
is a small image 718x791 in UTM Zone 18 at approx. \code{c(-77.76, 24.56)} from the
'lnglat' property of the open data set.
}
\examples{
## path to a 4-band TIFF image file
rgba_tif()
matrix(open.py(rgba_tif())$lnglat(), ncol = 2L,
         dimnames = list(NULL, c("longitude", "latitude")))
}
