% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{warp.py}
\alias{warp.py}
\title{Title}
\usage{
warp.py(
  fp,
  extent = c(-1e+05, 1e+05, -1e+05, 1e+05),
  dimension = c(100, 100),
  projection = "+proj=laea",
  resample = "nearest"
)
}
\arguments{
\item{fp}{gdal data source name (file, url, VRT string, database connection string, etc)}

\item{extent}{xmin,xmax,ymin,ymax}

\item{dimension}{width,height in pixels}

\item{projection}{projection string of output (interpreted by GDAL input reader)}
}
\value{
WarpedVRT dataset
}
\description{
Title
}
\examples{
warped <- warp.py(rgba_tif(), extent = c(-1e3, 1e3, -1e3, 1e3), dimension = c(256, 256),
              projection = sprintf("+proj=omerc +lon_0=\%f +lat_0=\%f +datum=WGS84",
                                                             -77.75791, 24.56158))
plot(warped)

longlat <- warp.py(rgba_tif(), extent =
                                1.2 * c(-1, 1, -1, 1) +
                               rep(c(-77.75791, 24.56158), each = 2L),
                              projection = "OGC:CRS84",
                              dimension = c(256, 256), resample = "cubic")
plot(longlat, asp = 1/cos(24.5*pi/180))
maps::map(add = TRUE, col = "firebrick", lwd = 4)
}
